requirements.txt
# Web Framework
flask==3.0.0
flask-sock==0.7.0
flask-cors==4.0.0

# OpenAI SDK
openai>=1.54.0

# WebSockets
websockets==12.0

# Twilio
twilio==8.10.0

# Audio
audioop-lts

# HTTP Client
aiohttp==3.9.1

# Environment
python-dotenv==1.0.0
